moved diff code into renderer ftm
